+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.6.0 - In Development
 
-...nothing here yet!
+* [#113](https://github.com/adamcaudill/yawast/issues/113) - Better False Positive Detection For Directory Search
 
 ## 0.5.2 - 2017-07-13