+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.6.0 - In Development
 
+* [#109](https://github.com/adamcaudill/yawast/issues/109) - DNS CAA Support
 * [#113](https://github.com/adamcaudill/yawast/issues/113) - Better False Positive Detection For Directory Search
 
 ## 0.5.2 - 2017-07-13