+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [#109](https://github.com/adamcaudill/yawast/issues/109) - DNS CAA Support
 * [#113](https://github.com/adamcaudill/yawast/issues/113) - Better False Positive Detection For Directory Search
+* [#115](https://github.com/adamcaudill/yawast/issues/115) - Add dns Command
 
 ## 0.5.2 - 2017-07-13