+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * [#115](https://github.com/adamcaudill/yawast/issues/115) - Add dns Command
 * [#116](https://github.com/adamcaudill/yawast/issues/116) - Add option '--nodns' to skip DNS checks
 * [#117](https://github.com/adamcaudill/yawast/issues/117) - Show additional information about the TLS connection
- 
+* [#118](https://github.com/adamcaudill/yawast/issues/118) - Add check for CVE-2017-12617 - Apache Tomcat PUT RCE
+
 ## 0.5.2 - 2017-07-13
 
 * [#107](https://github.com/adamcaudill/yawast/issues/107) - Current version check